activation based on purchase

+++ database/migrations/2020_05_08_191148_create_credit_requests_table.php
@@ -16,12 +16,12 @@ class CreateCreditRequestsTable extends Migration
         Schema::create('credit_requests', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->decimal('amount', 8, 2)->default(0);
-            $table->integer('requested_by');
+            $table->bigInteger('requested_by');
             $table->integer('payment_mode');
             $table->string('reference',512)->nullable();
             $table->integer('bank_id')->nullable();
             $table->string('status',32)->nullable();
-            $table->integer('approved_by');
+            $table->bigInteger('approved_by');
             $table->timestamp('approved_at')->nullable();
             $table->string('note',2048)->nullable();
             $table->timestamps();
